Fix test for existence of command
`which` will print, even if the command does not exist, so the test will
never fail. Additionally, `which` is non-standard, and has been
deprecated by debian. `command -v` is the POSIX-standardized way of
testing for existence of commands, and works the same even on non-POSIX
systems that implement the `command` command.

`command -v` will exit with 0 and print to stdout when the command
exists, and will exit with 1 and *not* print anything to stdout or
stderr when the command does not exist.